fix dsl style interface
